neutron: Move helpers into separate module
These helpers are used by multiple parts of the test
infrastructure.
