+++ zaza/openstack/charm_tests/neutron/utils.py
@@ -1,0 +1,74 @@
+# Copyright 2022 Canonical Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Helpers for neutron tests."""
+
+
+def floating_ips_from_instance(instance):
+    """
+    Retrieve floating IPs from an instance.
+
+    :param instance: The instance to fetch floating IPs from
+    :type instance: nova_client.Server
+
+    :returns: A list of floating IPs for the specified server
+    :rtype: list[str]
+    """
+    return ips_from_instance(instance, 'floating')
+
+
+def fixed_ips_from_instance(instance):
+    """
+    Retrieve fixed IPs from an instance.
+
+    :param instance: The instance to fetch fixed IPs from
+    :type instance: nova_client.Server
+
+    :returns: A list of fixed IPs for the specified server
+    :rtype: list[str]
+    """
+    return ips_from_instance(instance, 'fixed')
+
+
+def network_name_from_instance(instance):
+    """Retrieve name of primary network the instance is attached to.
+
+    :param instance: The instance to fetch name of network from.
+    :type instance: nova_client.Server
+    :returns: Name of primary network the instance is attached to.
+    :rtype: str
+    """
+    return next(iter(instance.addresses))
+
+
+def ips_from_instance(instance, ip_type):
+    """
+    Retrieve IPs of a certain type from an instance.
+
+    :param instance: The instance to fetch IPs from
+    :type instance: nova_client.Server
+    :param ip_type: the type of IP to fetch, floating or fixed
+    :type ip_type: str
+
+    :returns: A list of IPs for the specified server
+    :rtype: list[str]
+    """
+    if ip_type not in ['floating', 'fixed']:
+        raise RuntimeError(
+            "Only 'floating' and 'fixed' are valid IP types to search for"
+        )
+    return list([
+        ip['addr'] for ip in instance.addresses[
+            network_name_from_instance(instance)]
+        if ip['OS-EXT-IPS:type'] == ip_type])
